+++ sw/device/lib/crypto/README.md
@@ -1,5 +1,7 @@
 # Library of Cryptographic Implementations
 
-This folder contains implementations of various cryptographic algorithms. Each
-algorithm has its own, self-contained subfolder, and can be used independently
-of the others. See algorithms' individual READMEs for details.
+This folder contains OpenTitan's OS-visible cryptographic library. The public
+API is under include/; everything else should be considered internal.
+
+Tests are separately stored under sw/device/tests/crypto, except for driver
+tests which are under drivers/.
