+++ sw/device/lib/crypto/impl/rsa_3072/BUILD
@@ -15,9 +15,9 @@ cc_library(
         "//sw/device/lib/base:hardened",
         "//sw/device/lib/base:macros",
         "//sw/device/lib/base:memory",
-        "//sw/device/lib/crypto:otbn_util",
         "//sw/device/lib/crypto/drivers:hmac",
         "//sw/device/lib/crypto/drivers:otbn",
+        "//sw/device/lib/crypto/impl:otbn_util",
         "//sw/otbn/crypto:run_rsa_verify_3072",
     ],
 )
