+++ sw/device/lib/crypto/impl/rsa_3072/rsa_3072_verify.h
@@ -2,15 +2,15 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef OPENTITAN_SW_DEVICE_LIB_CRYPTO_RSA_3072_RSA_3072_VERIFY_H_
-#define OPENTITAN_SW_DEVICE_LIB_CRYPTO_RSA_3072_RSA_3072_VERIFY_H_
+#ifndef OPENTITAN_SW_DEVICE_LIB_CRYPTO_IMPL_RSA_3072_RSA_3072_VERIFY_H_
+#define OPENTITAN_SW_DEVICE_LIB_CRYPTO_IMPL_RSA_3072_RSA_3072_VERIFY_H_
 
 #include <stddef.h>
 #include <stdint.h>
 
 #include "sw/device/lib/base/hardened.h"
 #include "sw/device/lib/crypto/drivers/hmac.h"
-#include "sw/device/lib/crypto/otbn_util.h"
+#include "sw/device/lib/crypto/impl/otbn_util.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -129,4 +129,4 @@ otbn_error_t rsa_3072_verify(const rsa_3072_int_t *signature,
 }  // extern "C"
 #endif  // __cplusplus
 
-#endif  // OPENTITAN_SW_DEVICE_LIB_CRYPTO_RSA_3072_RSA_3072_VERIFY_H_
+#endif  // OPENTITAN_SW_DEVICE_LIB_CRYPTO_IMPL_RSA_3072_RSA_3072_VERIFY_H_
