+++ sw/device/tests/crypto/BUILD
@@ -18,7 +18,7 @@ opentitan_functest(
         timeout = "long",
     ),
     deps = [
-        "//sw/device/lib/crypto/aes_gcm",
+        "//sw/device/lib/crypto/impl/aes_gcm",
         "//sw/device/lib/runtime:log",
         "//sw/device/lib/testing/test_framework:check",
         "//sw/device/lib/testing/test_framework:ottf_main",
@@ -32,10 +32,10 @@ opentitan_functest(
         timeout = "long",
     ),
     deps = [
-        "//sw/device/lib/crypto:otbn_util",
         "//sw/device/lib/crypto/drivers:hmac",
         "//sw/device/lib/crypto/drivers:otbn",
-        "//sw/device/lib/crypto/ecdsa_p256",
+        "//sw/device/lib/crypto/impl:otbn_util",
+        "//sw/device/lib/crypto/impl/ecdsa_p256",
         "//sw/device/lib/runtime:log",
         "//sw/device/lib/testing:entropy_testutils",
         "//sw/device/lib/testing/test_framework:ottf_main",
@@ -57,10 +57,10 @@ opentitan_functest(
     ),
     deps = [
         ":ecdsa_p256_verify_testvectors_hardcoded_header",
-        "//sw/device/lib/crypto:otbn_util",
         "//sw/device/lib/crypto/drivers:hmac",
         "//sw/device/lib/crypto/drivers:otbn",
-        "//sw/device/lib/crypto/ecdsa_p256",
+        "//sw/device/lib/crypto/impl:otbn_util",
+        "//sw/device/lib/crypto/impl/ecdsa_p256",
         "//sw/device/lib/runtime:log",
         "//sw/device/lib/testing/test_framework:ottf_main",
     ],
@@ -81,10 +81,10 @@ opentitan_functest(
     ),
     deps = [
         ":rsa_3072_verify_testvectors_wycheproof_header",
-        "//sw/device/lib/crypto:otbn_util",
         "//sw/device/lib/crypto/drivers:hmac",
         "//sw/device/lib/crypto/drivers:otbn",
-        "//sw/device/lib/crypto/rsa_3072:rsa_3072_verify",
+        "//sw/device/lib/crypto/impl:otbn_util",
+        "//sw/device/lib/crypto/impl/rsa_3072:rsa_3072_verify",
         "//sw/device/lib/runtime:log",
         "//sw/device/lib/testing/test_framework:ottf_main",
     ],
@@ -105,10 +105,10 @@ opentitan_functest(
     ),
     deps = [
         ":rsa_3072_verify_testvectors_hardcoded_header",
-        "//sw/device/lib/crypto:otbn_util",
         "//sw/device/lib/crypto/drivers:hmac",
         "//sw/device/lib/crypto/drivers:otbn",
-        "//sw/device/lib/crypto/rsa_3072:rsa_3072_verify",
+        "//sw/device/lib/crypto/impl:otbn_util",
+        "//sw/device/lib/crypto/impl/rsa_3072:rsa_3072_verify",
         "//sw/device/lib/runtime:log",
         "//sw/device/lib/testing/test_framework:ottf_main",
     ],
